+++ tests/testthat/test-0-CRAN.R
@@ -48,4 +48,38 @@ test_that("Basic functionality works", {
       "\033[0;31mThere were another 8 warnings. To see all warnings call 'check_RoBTT(fit)'.\033[0m"
   )
   )
+  
+  # test setup
+  expect_equal(
+    capture_output_lines(check_setup(models = FALSE), print = TRUE, width = 150),
+    c("Robust Bayesian t-test (set-up)" ,
+      "Components summary:"             ,
+      "              Models Prior prob.",
+      "Effect           4/8       0.500",
+      "Heterogeneity    4/8       0.500",
+      "Outliers         4/8       0.500"
+    )
+  )
+  
+  expect_equal(
+    capture_output_lines(check_setup(models = TRUE), print = TRUE, width = 150),
+    c("Robust Bayesian t-test (set-up)"                                          ,
+      "Components summary:"                                                      ,
+      "              Models Prior prob."                                         ,
+      "Effect           4/8       0.500"                                         ,
+      "Heterogeneity    4/8       0.500"                                         ,
+      "Outliers         4/8       0.500"                                         ,
+      ""                                                                         ,
+      "Models overview:"                                                         ,
+      " Model Distribution   Prior delta    Prior rho    Prior nu    Prior prob.",
+      "     1       normal        Spike(0) Spike(0.5)                      0.125",
+      "     2            t        Spike(0) Spike(0.5) Exponential(1)       0.125",
+      "     3       normal        Spike(0) Beta(1, 1)                      0.125",
+      "     4            t        Spike(0) Beta(1, 1) Exponential(1)       0.125",
+      "     5       normal Cauchy(0, 0.71) Spike(0.5)                      0.125",
+      "     6            t Cauchy(0, 0.71) Spike(0.5) Exponential(1)       0.125",
+      "     7       normal Cauchy(0, 0.71) Beta(1, 1)                      0.125",
+      "     8            t Cauchy(0, 0.71) Beta(1, 1) Exponential(1)       0.125"
+    )
+  )
 })