+++ NEWS.md
@@ -1,3 +1,9 @@
+## version 1.1.0
+### Features
+- `check_setup()` function
+- `diagnostics()` function
+- `rho2logsdr` transformations
+
 ## version 1.0.3
 Compatibility update for the rstan 2.31.
 (correct CRAN commit)